add update method
This commit adds an endpoint for setting the word of the day. Simply
put, if a user of the API receives an error for the current word of the
day, they may use this endpoint to pick a new word. They can then retry
the word of the day endpoint to retrieve the new word.